FMWK-788 Update to new bandwidth limiter (#409)
* FMWK-788-new-limiter
- testing limiter

* origin/FMWK-788-new-limiter
- test default values

* FMWK-788-new-limiter
- removed limit

* FMWK-788-new-limiter
- increase test records

* FMWK-788-new-limiter
- remove limit

* FMWK-788-new-limiter
- fix

* FMWK-788-new-limiter
- new validation
- int64 for bandwidth

* FMWK-788-new-limiter
- removed unused casting

* FMWK-788-new-limiter
- linter
- tests

* FMWK-788-new-limiter
- lib update

* use errValidationNegative

* regen doc

---------

+++ pkg/dto/backup_policy_test.go
@@ -15,7 +15,7 @@ func TestBackupPolicyConversionIsLossless(t *testing.T) {
 	noRecords := true
 	noIndexes := false
 	noUdfs := true
-	bandwidth := 50
+	bandwidth := int64(50)
 	recordsPerSecond := 100
 	fileLimit := 1024
 	compressionPolicy := &CompressionPolicy{Level: 5}
